+++ spec/abstract.spec.ts
@@ -1,10 +1,13 @@
-import { BuiltinClass } from '../src';
+import { Es5BuiltinClass, Es6BuiltinClass } from '../src';
+import { TARGET } from './helper';
+
+const BuiltinClass = TARGET == 'es5' ? Es5BuiltinClass : Es6BuiltinClass;
 
 @BuiltinClass()
 abstract class Abstract extends Error {}
 class Concreate extends Abstract {}
 
-it('nested', () => {
+it('abstract', () => {
   expect(new Concreate() instanceof Concreate).toBe(true);
   expect(new Concreate() instanceof Abstract).toBe(true);
   expect(new Concreate() instanceof Error).toBe(true);