+++ spec/nested.spec.ts
@@ -1,5 +1,7 @@
-// tslint:disable:max-classes-per-file
-import { BuiltinClass } from '../src';
+import { Es5BuiltinClass, Es6BuiltinClass } from '../src';
+import { TARGET } from './helper';
+
+const BuiltinClass = TARGET == 'es5' ? Es5BuiltinClass : Es6BuiltinClass;
 
 @BuiltinClass()
 class E1 extends Error {}