+++ spec/unneeded.spec.ts
@@ -1,8 +1,0 @@
-import { UNNEEDED } from '../src';
-
-class E extends Error {}
-const EXPECTED = new E() instanceof E;
-
-it('UNNEEDED', () => {
-  expect(UNNEEDED).toBe(EXPECTED);
-});