+++ src/index.ts
@@ -1,24 +1,44 @@
-class E extends Error {}
-export const UNNEEDED = new E() instanceof E;
-
 // eslint-disable-next-line @typescript-eslint/ban-types,@typescript-eslint/no-explicit-any
-export type AbstractConstructorType = Function & { prototype: any };
+type AbstractConstructorType = Function & { prototype: any };
 // eslint-disable-next-line @typescript-eslint/ban-types,@typescript-eslint/no-explicit-any
-export type ConcreteConstructorType = new (...args: any[]) => {};
-export type ConstructorType = AbstractConstructorType | ConcreteConstructorType;
+type ConcreteConstructorType = new (...args: any[]) => {};
+type ConstructorType = AbstractConstructorType | ConcreteConstructorType;
 
-export function BuiltinClass() {
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-  return function <T extends ConstructorType>(constructor: T) {
-    if (UNNEEDED) {
-      return constructor;
-    }
-    return class extends (constructor as ConcreteConstructorType) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      constructor(...args: any[]) {
-        super(...args);
-        Object.setPrototypeOf(this, new.target.prototype);
-      }
-    } as T;
-  };
+type BuiltinClassType = () => <T extends ConstructorType>(constructor: T) => T;
+
+class Target {
+  constructor(private _builtinClass?: BuiltinClassType) {}
+
+  isSupported() {
+    return !!this._builtinClass;
+  }
+
+  get builtinClass() {
+    return this.isSupported()
+      ? (this._builtinClass as BuiltinClassType)
+      : function () {
+          return function <T extends ConstructorType>(constructor: T): T {
+            return constructor;
+          };
+        };
+  }
 }
+
+function loadTarget(target: string) {
+  try {
+    //eslint-disable-next-line @typescript-eslint/no-var-requires,@typescript-eslint/no-unsafe-member-access
+    return new Target(require(`./${target}`).BuiltinClass as BuiltinClassType);
+  } catch (e) {
+    if (e instanceof SyntaxError) {
+      return new Target();
+    }
+    throw e;
+  }
+}
+
+const es5 = loadTarget('es5');
+const es6 = loadTarget('es6');
+
+export const Es5BuiltinClass = es5.builtinClass;
+export const Es6BuiltinClass = es6.builtinClass;
+export const ES6_SUPPORTED = es6.isSupported();