+++ CHANGELOG.md
@@ -7,6 +7,16 @@ Based on [Keep a Changelog] and [Semantic Versioning Caret].
 
 * Nothing yet.
 
+## [0.4.0]
+
+### Added
+
+* Added the `Es5BuiltinClass` and `Es6BuiltinClass` decorators for improving es5/es6 compatibility. 
+
+### Removed
+
+* Removed the `BuiltinClass` decorator. Use `Es5BuiltinClass`/`Es6BuiltinClass` instead.
+
 ## [0.3.1]
 
 ### Fixed
@@ -33,7 +43,9 @@ Based on [Keep a Changelog] and [Semantic Versioning Caret].
 
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning Caret]: https://github.com/myowncake/semver-caret
-[edge]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.3.0...HEAD
+[edge]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.3.1...v0.3.2
+[0.3.1]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/typescript-plus/builtin-class-decorator/compare/v0.1.0...v0.1.1